+++ library/src/main/java/com/feizhang/validation/annotations/Min.java
@@ -8,7 +8,7 @@ import java.lang.annotation.Target;
 @Target({ElementType.FIELD})
 @Retention(RetentionPolicy.RUNTIME)
 public @interface Min {
-    int min();
+    long min();
 
     String message() default "%s不能小于%d";
 }
